+++ crates/migration/src/lib.rs
@@ -34,7 +34,7 @@ mod m20260616_000013_rename_cycle_defaulted_to_settling;
 mod m20260618_000014_collateral_event_credit_default;
 mod m20260629_000015_settlement_cycle_status_shortfall;
 mod m20260707_000016_blockchain_event_dead_letter;
-mod m20260707_000016_settlement_cycle_status_confirmed;
+mod m20260707_000017_settlement_cycle_status_confirmed;
 
 pub struct Migrator;
 
@@ -76,7 +76,7 @@ impl MigratorTrait for Migrator {
             Box::new(m20260618_000014_collateral_event_credit_default::Migration),
             Box::new(m20260629_000015_settlement_cycle_status_shortfall::Migration),
             Box::new(m20260707_000016_blockchain_event_dead_letter::Migration),
-            Box::new(m20260707_000016_settlement_cycle_status_confirmed::Migration),
+            Box::new(m20260707_000017_settlement_cycle_status_confirmed::Migration),
         ]
     }
 }
